[TC] Regenerated demo for tc_tc_compare_mode_sam_e70_xult
MH3-67829

+++ apps/tc/tc_compare_mode/firmware/src/config/sam_e70_xult/peripheral/tc/plib_tc3.c
@@ -118,7 +118,7 @@ void TC3_CH2_ComparePeriodSet (uint16_t period)
 /* Read the period value */
 uint16_t TC3_CH2_ComparePeriodGet (void)
 {
-    return TC3_REGS->TC_CHANNEL[2].TC_RC;
+    return (uint16_t)TC3_REGS->TC_CHANNEL[2].TC_RC;
 }
 
 /* Set the compare A value */
